face-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
insightface==0.7.3
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
numpy>=1.24.0,<2.0.0
python-multipart==0.0.18
Pillow==11.1.0
